refactor: move BaseExtractor to core module for better architecture
- Moved MetadataExtractor from metadata/core to core module
- Updated all imports across the codebase to use new location
- Fixed circular import issues using TYPE_CHECKING
- Moved metadata_types.py to metadata/types.py for cleaner structure
- Updated tests to reflect new import paths and attribute names
- All base classes (BaseMSIReader, BaseMSIConverter, MetadataExtractor)
  now properly reside in the core module

This change improves the architectural consistency by keeping all
abstract base classes together in the core module.
