+++ msiconvert/core/__init__.py
@@ -1,4 +1,5 @@
 # msiconvert/core/__init__.py
 from .base_converter import BaseMSIConverter
+from .base_extractor import MetadataExtractor
 from .base_reader import BaseMSIReader
 from .registry import register_converter, register_format_detector, register_reader
