--- msiconvert/metadata/core/__init__.py
@@ -1,1 +1,0 @@
-# msiconvert/metadata/core/__init__.py
