+++ msiconvert/metadata/types.py
